+++ src/Providers/None.php
@@ -99,6 +99,18 @@ class None implements Environment
     /**
      * {@inheritdoc}
      */
+    public function getPreviousCommit()
+    {
+        if (!$this->isGitRepo()) {
+            return '';
+        }
+
+        return exec('git log --pretty=format:"%H" -1 --skip=1');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function getAuthor()
     {
         if (!$this->isGitRepo()) {